Moved RestControllers into rest directory.
Created FileController
